scraper/requirements.txt
aiohttp>=3.9.5
requests>=2.31.0
supabase>=2.7.0
python-dotenv>=1.0.1
pytest>=8.1.1
pytest-asyncio>=0.23.6
aioresponses>=0.7.6
